+++ src/components/calcite-radio-group-item/calcite-radio-group-item.tsx
@@ -32,10 +32,15 @@ export class CalciteRadioGroupItem {
   //
   //--------------------------------------------------------------------------
 
-  /**
-   * Indicates whether the control is checked.
-   */
+  /** Indicates whether the control is checked. */
   @Prop({ reflect: true, mutable: true }) checked = false;
+
+  /** optionally pass an icon to display - accepts Calcite UI icon names  */
+  @Prop({ reflect: true }) icon?: string;
+
+  /** optionally used with icon, select where to position the icon */
+  @Prop({ reflect: true, mutable: true }) iconPosition?: "start" | "end" =
+    "start";
 
   @Watch("checked")
   protected handleCheckedChange(): void {
@@ -69,6 +74,11 @@ export class CalciteRadioGroupItem {
     }
 
     this.inputProxy = inputProxy;
+
+    // prop validations
+    let iconPosition = ["start", "end"];
+    if (this.icon !== null && !iconPosition.includes(this.iconPosition))
+      this.iconPosition = "start";
   }
 
   disconnectedCallback() {
@@ -86,6 +96,11 @@ export class CalciteRadioGroupItem {
     const { checked, useFallback, value } = this;
     const scale = getElementProp(this.el, "scale", "m");
     const appearance = getElementProp(this.el, "appearance", "m");
+    const layout = getElementProp(this.el, "layout", "m");
+
+    const iconEl = (
+      <calcite-icon class="radio-group-item-icon" icon={this.icon} scale="s" />
+    );
 
     return (
       <Host
@@ -93,10 +108,13 @@ export class CalciteRadioGroupItem {
         aria-checked={checked.toString()}
         scale={scale}
         appearance={appearance}
+        layout={layout}
       >
         <label>
+          {this.icon && this.iconPosition === "start" ? iconEl : null}
           <slot>{useFallback ? value : ""}</slot>
           <slot name="input" />
+          {this.icon && this.iconPosition === "end" ? iconEl : null}
         </label>
       </Host>
     );